feat: introduce Element component
