+++ apps/ui/src/app/[locale]/dev/pages-overview/page.tsx
@@ -2,7 +2,7 @@ import type { Locale } from "next-intl"
 import { setRequestLocale } from "next-intl/server"
 
 import PageList from "@/app/[locale]/dev/pages-overview/components/PageList"
-import Typography from "@/components/typography"
+import Typography from "@/components/html/components/Typography"
 import { logNonBlockingError } from "@/lib/logging"
 import { PublicStrapiClient } from "@/lib/strapi-api"
